# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TikTok Dropship AI Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all saved products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Save a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Save a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        estimatedCost:
          type: ["number", "null"]
        estimatedSellingPrice:
          type: ["number", "null"]
        profitMargin:
          type: ["number", "null"]
        trendScore:
          type: ["integer", "null"]
        status:
          type: string
          enum: [researching, active, paused, archived]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - category
        - status
        - createdAt
        - updatedAt
    ProductInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
        estimatedCost:
          type: ["number", "null"]
        estimatedSellingPrice:
          type: ["number", "null"]
        trendScore:
          type: ["integer", "null"]
        status:
          type: string
          enum: [researching, active, paused, archived]
        notes:
          type: ["string", "null"]
      required:
        - name
        - description
        - category
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        estimatedCost:
          type: ["number", "null"]
        estimatedSellingPrice:
          type: ["number", "null"]
        trendScore:
          type: ["integer", "null"]
        status:
          type: string
          enum: [researching, active, paused, archived]
        notes:
          type: ["string", "null"]
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        platform:
          type: string
        url:
          type: ["string", "null"]
        productCategory:
          type: string
        rating:
          type: ["number", "null"]
        minOrderQuantity:
          type: ["integer", "null"]
        shippingTime:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - platform
        - productCategory
        - createdAt
    SupplierInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        platform:
          type: string
        url:
          type: ["string", "null"]
        productCategory:
          type: string
        rating:
          type: ["number", "null"]
        minOrderQuantity:
          type: ["integer", "null"]
        shippingTime:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
        - platform
        - productCategory
...