# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TikTok Dropship AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from "./productStatus";

export interface Product {
  id: number;
  name: string;
  description: string;
  category: string;
  /** @nullable */
  estimatedCost?: number | null;
  /** @nullable */
  estimatedSellingPrice?: number | null;
  /** @nullable */
  profitMargin?: number | null;
  /** @nullable */
  trendScore?: number | null;
  status: ProductStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}