# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TikTok Dropship AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputStatus } from "./productInputStatus";

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  category: string;
  /** @nullable */
  estimatedCost?: number | null;
  /** @nullable */
  estimatedSellingPrice?: number | null;
  /** @nullable */
  trendScore?: number | null;
  status?: ProductInputStatus;
  /** @nullable */
  notes?: string | null;
}